+++ union/rand.h
@@ -38,70 +38,70 @@ int intRand(int start, int finish)
 
 struct PointF
 {
-    double x;
-    double y;
+    long double x;
+    long double y;
 };
 
 struct Delta
 {
-    double deltaX;
-    double deltaY;
+    long double deltaX;
+    long double deltaY;
 };
 
 
 Delta getDeltas(PointF &a, PointF &b)
 {
-    double deltaX = b.x - a.x;
-    double deltaY = b.y - a.y;
+    long double deltaX = b.x - a.x;
+    long double deltaY = b.y - a.y;
     return {deltaX, deltaY};
 }
 
 Delta getDeltas(PointF *a, PointF *b)
 {
-    double deltaX = b->x - a->x;
-    double deltaY = b->y - a->y;
+    long double deltaX = b->x - a->x;
+    long double deltaY = b->y - a->y;
     return {deltaX, deltaY};
 }
 
 
 
 
-double getDis(Delta &del)
+long double getDis(Delta &del)
 {
-    double c = sqrt(del.deltaX * del.deltaX + del.deltaY * del.deltaY);
+    long double c = sqrt(del.deltaX * del.deltaX + del.deltaY * del.deltaY);
     return c;
 };
 
 
-double getDis(Delta *del)
+long double getDis(Delta *del)
 {
-    double c = sqrt(del->deltaX * del->deltaX + del->deltaY * del->deltaY);
+    long double c = sqrt(del->deltaX * del->deltaX + del->deltaY * del->deltaY);
     return c;
 };
 
 
 
-double getConor(Delta &del)
+long double getConor(Delta &del)
 {
-    double conorData = atan2(del.deltaY, del.deltaX);
+    long double conorData = atan2(del.deltaY, del.deltaX);
     return conorData;
 };
 
 
 
-double getConor(Delta *del)
+long double getConor(Delta *del)
 {
-    double conorData = atan2(del->deltaY, del->deltaX);
+    long double conorData = atan2(del->deltaY, del->deltaX);
     return conorData;
 };
 
 
-double radToDeg(double rad)
+long double radToDeg(long double rad)
 {
     return rad * 180 / M_PI;
 }
 
-double degToRad(double deg)
+long double degToRad(long double deg)
 {
     return deg * (M_PI / 180);
 }
